+++ src/token/transport.go
@@ -1,87 +1,149 @@
 package token
 
 import (
+	"bytes"
 	"context"
 	"encoding/json"
+	"errors"
 	"io/ioutil"
 	"net/http"
 
 	"github.com/gorilla/mux"
 )
 
-type TokenRequestBuilder func(*http.Request, *Request) error
+var (
+	ErrNoHttp = errors.New("At least one of header, parameter, or variable must be specified")
+)
 
-func HeaderClaim(header, claim string) TokenRequestBuilder {
-	return func(hr *http.Request, tr *Request) error {
-		v := hr.Header.Get(header)
-		if len(v) > 0 {
-			tr.Claims[claim] = v
-		}
+type RequestBuilder func(*http.Request, *Request) error
 
-		return nil
-	}
+type missingValueError struct {
+	name  string
+	value *Value
 }
 
-func ParameterClaim(parameter, claim string) TokenRequestBuilder {
-	return func(hr *http.Request, tr *Request) error {
-		v := hr.Form.Get(parameter)
-		if len(v) > 0 {
-			tr.Claims[claim] = v
+func (mve *missingValueError) Error() string {
+	var output bytes.Buffer
+	output.WriteString("Missing value '")
+	output.WriteString(mve.name)
+	output.WriteString("' from")
+
+	first := true
+	if len(mve.value.Header) > 0 {
+		output.WriteString(" header '")
+		output.WriteString(mve.value.Header)
+		output.WriteString("'")
+		first = false
+	}
+
+	if len(mve.value.Parameter) > 0 {
+		if !first {
+			output.WriteString(" or")
+		}
+
+		output.WriteString(" parameter '")
+		output.WriteString(mve.value.Parameter)
+		output.WriteString("'")
+		first = false
+	}
+
+	if len(mve.value.Variable) > 0 {
+		if !first {
+			output.WriteString(" or")
+		}
+
+		output.WriteString(" variable '")
+		output.WriteString(mve.value.Variable)
+		output.WriteString("'")
+		first = false
+	}
+
+	return output.String()
+}
+
+func (mve *missingValueError) StatusCode() int {
+	return http.StatusBadRequest
+}
+
+func extractValue(original *http.Request, value *Value) (interface{}, bool) {
+	if len(value.Header) > 0 {
+		hv := original.Header.Get(value.Header)
+		if len(hv) > 0 {
+			return hv, true
+		}
+	}
+
+	if len(value.Parameter) > 0 {
+		pv := original.Form[value.Parameter]
+		if len(pv) > 0 {
+			return pv[0], true
+		}
+	}
+
+	if len(value.Variable) > 0 {
+		vv := mux.Vars(original)[value.Variable]
+		if len(vv) > 0 {
+			return vv, true
+		}
+	}
+
+	return nil, false
+}
+
+func NewClaimBuilder(name string, value Value) (RequestBuilder, error) {
+	if !value.IsHttp() {
+		return nil, ErrNoHttp
+	}
+
+	mve := missingValueError{name: name, value: &value}
+	return func(original *http.Request, tr *Request) error {
+		if v, ok := extractValue(original, &value); ok {
+			tr.Claims[name] = v
 			return nil
 		}
 
-		v = hr.PostForm.Get(parameter)
-		if len(v) > 0 {
-			tr.Claims[claim] = v
-		}
-
-		return nil
-	}
+		return &mve
+	}, nil
 }
 
-func VariableClaim(variable, claim string) TokenRequestBuilder {
-	return func(hr *http.Request, tr *Request) error {
-		v := mux.Vars(hr)[variable]
-		if len(v) > 0 {
-			tr.Claims[claim] = v
-		}
-
-		return nil
+func NewMetadataBuilder(name string, value Value) (RequestBuilder, error) {
+	if !value.IsHttp() {
+		return nil, ErrNoHttp
 	}
-}
 
-func MetaHeader(header, key string) TokenRequestBuilder {
-	return func(hr *http.Request, tr *Request) error {
-		v := hr.Header.Get(header)
-		if len(v) > 0 {
-			tr.Meta[key] = v
-		}
-
-		return nil
-	}
-}
-
-func MetaParameter(parameter, key string) TokenRequestBuilder {
-	return func(hr *http.Request, tr *Request) error {
-		v := hr.Form.Get(parameter)
-		if len(v) > 0 {
-			tr.Meta[key] = v
+	mve := missingValueError{name: name, value: &value}
+	return func(original *http.Request, tr *Request) error {
+		if v, ok := extractValue(original, &value); ok {
+			tr.Metadata[name] = v
 			return nil
 		}
 
-		v = hr.PostForm.Get(parameter)
-		if len(v) > 0 {
-			tr.Meta[key] = v
-		}
-
-		return nil
-	}
+		return &mve
+	}, nil
 }
 
-func BuildRequest(hr *http.Request, b ...TokenRequestBuilder) (*Request, error) {
+func NewTokenRequestBuilders(o Options) []RequestBuilder {
+	var builders []RequestBuilder
+
+	for name, value := range o.Claims {
+		if b, err := NewClaimBuilder(name, value); err == nil {
+			builders = append(builders, b)
+		}
+	}
+
+	for name, value := range o.Metadata {
+		if b, err := NewMetadataBuilder(name, value); err == nil {
+			builders = append(builders, b)
+		}
+	}
+
+	return builders
+}
+
+func BuildRequest(hr *http.Request, b []RequestBuilder) (*Request, error) {
 	tr := &Request{
-		Claims: make(map[string]interface{}),
-		Meta:   make(map[string]interface{}),
+		Claims:   make(map[string]interface{}),
+		Metadata: make(map[string]interface{}),
 	}
 
 	for _, f := range b {
@@ -93,30 +155,9 @@ func BuildRequest(hr *http.Request, b ...TokenRequestBuilder) (*Request, error) 
 	return tr, nil
 }
 
-func NewTokenRequestBuilders(d Descriptor) []TokenRequestBuilder {
-	var rbs []TokenRequestBuilder
-	for header, claim := range d.HeaderClaims {
-		rbs = append(rbs, HeaderClaim(header, claim))
-	}
-
-	for parameter, claim := range d.ParameterClaims {
-		rbs = append(rbs, ParameterClaim(parameter, claim))
-	}
-
-	for header, key := range d.MetaHeaders {
-		rbs = append(rbs, MetaHeader(header, key))
-	}
-
-	for parameter, key := range d.MetaParameters {
-		rbs = append(rbs, MetaParameter(parameter, key))
-	}
-
-	return rbs
-}
-
-func DecodeServerRequest(b ...TokenRequestBuilder) func(context.Context, *http.Request) (interface{}, error) {
+func DecodeServerRequest(b ...RequestBuilder) func(context.Context, *http.Request) (interface{}, error) {
 	return func(_ context.Context, hr *http.Request) (interface{}, error) {
-		tr, err := BuildRequest(hr, b...)
+		tr, err := BuildRequest(hr, b)
 		if err != nil {
 			return nil, err
 		}